backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
elevenlabs[pyaudio]==2.25.0
pinecone-client
openai
python-dotenv
